+++ src/components/dictionaries/zdic/View.tsx
@@ -1,22 +1,21 @@
 import React from 'react'
-import Speaker from '@/components/Speaker'
 import { ZdicResult } from './engine'
 import { ViewPorps } from '@/components/dictionaries/helpers'
+import { withStaticSpeaker } from '@/components/withStaticSpeaker'
+import EntryBox from '@/components/EntryBox'
 
-export default class DictZdic extends React.PureComponent<ViewPorps<ZdicResult>> {
+export class DictZdic extends React.PureComponent<ViewPorps<ZdicResult>> {
   render () {
-    const { defs, phsym } = this.props.result
     return (
-      <>
-        <ul className='dictZdic-Phsym'>
-          {phsym.map(p => (
-            <li key={p.pron} className='dictZdic-PhsymItem'>
-              {p.pinyin} <Speaker src={p.pron} />
-            </li>
-          ))}
-        </ul>
-        <div className='dictZdic-Defs' dangerouslySetInnerHTML={{ __html: defs }} />
-      </>
+      <div className='dictZidc-Entry'>
+        {this.props.result.map(entry => (
+          <EntryBox title={entry.title} key={entry.title}>
+            <div dangerouslySetInnerHTML={{ __html: entry.content }} />
+          </EntryBox>
+        ))}
+      </div>
     )
   }
 }
+
+export default withStaticSpeaker(DictZdic)