+++ src/content/components/DictPanel/index.tsx
@@ -52,14 +52,18 @@ export interface DictPanelProps extends ChildrenProps {
 
 interface DictPanelState {
   mtaBoxHeight: number
+  audioBoxShow: boolean
+  shouldLoadAudioBox: boolean
 }
 
 export class DictPanel extends React.Component<DictPanelProps & { t: TranslationFunction }, DictPanelState> {
   ContainerRef = React.createRef<HTMLMainElement>()
   MtaBoxRef = React.createRef<HTMLTextAreaElement>()
 
-  state = {
-    mtaBoxHeight: 0
+  state: DictPanelState = {
+    mtaBoxHeight: 0,
+    audioBoxShow: false,
+    shouldLoadAudioBox: false,
   }
 
   searchText = (arg?: { id?: DictID, info?: SelectionInfo | string }) => {
@@ -85,12 +89,7 @@ export class DictPanel extends React.Component<DictPanelProps & { t: Translation
 
   toggleMtaBox = (e?: React.MouseEvent<HTMLButtonElement>) => {
     if (e) { e.currentTarget.blur() }
-    this.setState(preState => {
-      return { mtaBoxHeight: preState.mtaBoxHeight <= 0
-        ? window.innerHeight * this.props.panelMaxHeightRatio / 100 * 0.4
-        : 0
-      }
-    })
+    this.showMtaBox(this.state.mtaBoxHeight <= 0)
   }
 
   handleMtaBoxInput = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
@@ -115,6 +114,22 @@ export class DictPanel extends React.Component<DictPanelProps & { t: Translation
     }
   }
 
+  showAudioBox = (isShow: boolean) => {
+    this.setState({ audioBoxShow: isShow })
+  }
+
+  toggleAudioBox = (e?: React.MouseEvent<HTMLButtonElement>) => {
+    if (e) { e.currentTarget.blur() }
+    this.showAudioBox(!this.state.audioBoxShow)
+  }
+
+  handleAudioBoxInput = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    this.props.searchBoxUpdate({
+      index: this.props.searchBox.index,
+      text: e.currentTarget.value
+    })
+  }
+
   componentDidMount () {
     const { mtaAutoUnfold } = this.props
     if (mtaAutoUnfold === 'once' ||
@@ -129,6 +144,7 @@ export class DictPanel extends React.Component<DictPanelProps & { t: Translation
         this.MtaBoxRef.current.focus()
         this.MtaBoxRef.current.select()
       }
+      this.setState({ shouldLoadAudioBox: true })
     }, 100)
   }
 
@@ -202,6 +218,8 @@ export class DictPanel extends React.Component<DictPanelProps & { t: Translation
 
     const {
       mtaBoxHeight,
+      audioBoxShow,
+      shouldLoadAudioBox,
     } = this.state
 
     const {
@@ -245,30 +263,49 @@ export class DictPanel extends React.Component<DictPanelProps & { t: Translation
               {this.renderMtaBox}
             </CSSTransition>
           </div>
-          <button className='panel-MtaBoxBtn' onClick={this.toggleMtaBox}>
+          <button className='panel-DrawerBtn' onClick={this.toggleMtaBox}>
             <svg width='10' height='10' viewBox='0 0 59.414 59.414' xmlns='http://www.w3.org/2000/svg'
-             className={'panel-MtaBoxBtn_Arrow' + (mtaBoxHeight > 0 ? ' isActive' : '')}
+             className={'panel-DrawerBtn_Arrow' + (mtaBoxHeight > 0 ? ' isActive' : '')}
             >
               <path d='M58 14.146L29.707 42.44 1.414 14.145 0 15.56 29.707 45.27 59.414 15.56' />
             </svg>
           </button>
-          {activeDicts.map(id => {
-            const dictInfo = dictsInfo[id]
-            return React.createElement(DictItem, {
-              t,
-              key: id,
-              id,
-              text: (dictionaries.searchHistory[0] || selection.selectionInfo).text,
-              fontSize,
-              preferredHeight: dictsConfig.all[id].preferredHeight,
-              panelWidth,
-              searchStatus: dictInfo ? dictInfo.searchStatus : SearchStatus.OnHold,
-              searchResult: dictInfo ? dictInfo.searchResult : null,
-              searchText: this.searchText,
-              updateItemHeight,
-            })
-          })}
+          <div>
+            {activeDicts.map(id => {
+              const dictInfo = dictsInfo[id]
+              return React.createElement(DictItem, {
+                t,
+                key: id,
+                id,
+                text: (dictionaries.searchHistory[0] || selection.selectionInfo).text,
+                fontSize,
+                preferredHeight: dictsConfig.all[id].preferredHeight,
+                panelWidth,
+                searchStatus: dictInfo ? dictInfo.searchStatus : SearchStatus.OnHold,
+                searchResult: dictInfo ? dictInfo.searchResult : null,
+                searchText: this.searchText,
+                updateItemHeight,
+              })
+            })}
+          </div>
         </main>
+        <button className='panel-DrawerBtn' onClick={this.toggleAudioBox}>
+          <svg width='10' height='10' viewBox='0 0 59.414 59.414' xmlns='http://www.w3.org/2000/svg'
+            className={'panel-DrawerBtn_Arrow' + (audioBoxShow ? ' isActive' : '')}
+          >
+            <path d='M 58 45.269 L 29.707 16.975 L 1.414 45.27 L 0 43.855 L 29.707 14.145 L 59.414 43.855' />
+          </svg>
+        </button>
+        <div className='panel-AudioBox' style={{ height: audioBoxShow ? 165 : 0 }}>
+          {shouldLoadAudioBox // x-frame-options: SAMEORIGIN
+            ? <iframe
+                className='panel-AudioBoxFrame'
+                src={browser.runtime.getURL('/audio-control.html')}
+                sandbox='allow-same-origin allow-scripts'
+              />
+            : null
+          }
+        </div>
       </div>
     )
   }