+++ test/specs/background/server.spec.ts
@@ -21,6 +21,9 @@ describe('Server', () => {
   browserWrap.openURL = openURL
 
   beforeAll(() => {
+    browser.runtime.sendMessage.callsFake(() => Promise.resolve({}))
+    browser.tabs.sendMessage.callsFake(() => Promise.resolve({}))
+
     jest.doMock('@/background/sync-manager', () => {})
 
     jest.doMock('@/_helpers/chs-to-chz', () => {
@@ -73,6 +76,7 @@ describe('Server', () => {
 
   beforeEach(() => {
     browser.flush()
+    browser.runtime.sendMessage.callsFake(() => Promise.resolve({}))
     chsToChz.mockReset()
     chsToChz.mockImplementation(t => t)
     play.mockReset()
@@ -99,15 +103,6 @@ describe('Server', () => {
     expect(chsToChz).toHaveBeenCalledTimes(0)
     expect(openURL).toHaveBeenCalledTimes(1)
     expect(openURL).toHaveBeenCalledWith('https://test.com/', undefined)
-  })
-
-  it('Audio Play', () => {
-    browser.runtime.onMessage.dispatch({
-      type: MsgType.PlayAudio,
-      src: 'https://test.com/a.mp3',
-    })
-    expect(play).toHaveBeenCalledTimes(1)
-    expect(play).toHaveBeenCalledWith('https://test.com/a.mp3')
   })
 
   describe('Fetch Dict Result', () => {